src/requirements.txt
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
sqlalchemy
psycopg2-binary
requests 
graypy
pydantic_settings